fix: Extract AvatarComponent
